requirements.txt
aactivator==1.0.1
aiohttp==3.8.1
aiosignal==1.2.0
alpha-vantage==2.3.1
anyio==3.6.1
appnope==0.1.3
APScheduler==3.9.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bidict==0.22.0
bleach==5.0.0
Brotli==1.0.9
bs4==0.0.1
build==0.8.0
cachetools==5.0.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.0
click==8.1.3
configparser==5.2.0
coverage==6.4.3
cryptography==37.0.2
cycler==0.11.0
Cython==0.29.30
dash==2.6.1
dash-bootstrap-components==1.0.3
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-iconify==0.1.2
dash-table==5.0.0
dateparser==1.1.1
DateTime==4.4
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.4
distlib==0.3.5
distro==1.7.0
entrypoints==0.4
env-file==2020.12.3
eventkit==0.8.9
fastjsonschema==2.15.3
filelock==3.6.0
flake8==3.9.2
Flask==2.2.2
Flask-Compress==1.12
fonttools==4.34.4
formulaic==0.3.4
frozenlist==1.3.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
ib-insync==0.9.70
idna==3.3
imageio==2.19.1
importlib-metadata==4.12.0
importlib-resources==5.7.1
iniconfig==1.1.1
interface-meta==1.3.0
ipykernel==6.13.0
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.5.1
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kaggle==1.5.12
kiwisolver==1.4.4
linearmodels==4.25
llvmlite==0.36.0
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
multidict==6.0.2
multitasking==0.0.11
mypy==0.910
mypy-extensions==0.4.3
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.11
numba==0.53.1
numpy==1.21.6
outcome==1.2.0
packaging==21.3
pandas==1.3.5
pandas-datareader==0.10.0
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.2
pep517==0.13.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pipenv==2022.1.8
platformdirs==2.5.1
plotly==5.9.0
pluggy==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
property-cached==1.6.4
psutil==5.9.0
ptyprocess==0.7.0
py==1.11.0
pycodestyle==2.7.0
pycparser==2.21
pyflakes==2.3.1
Pygments==2.12.0
pyhdfe==0.1.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
pyTelegramBotAPI==4.6.0
pytest==6.2.5
pytest-cov==2.12.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-engineio==4.3.1
python-slugify==6.1.1
python-socketio==5.0.4
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==22.3.0
regex==2022.3.2
requests==2.28.1
rfc3986==1.5.0
schedule==1.1.0
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
selenium==4.3.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
socketIO-client==0.7.2
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.36
statsmodels==0.13.2
tabula-py==2.4.0
tabulate==0.8.10
tenacity==8.0.1
terminado==0.13.3
testpath==0.6.0
text-unidecode==1.3
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tornado==6.1
tox==3.24.3
tqdm==4.64.0
traitlets==5.3.0
trio==0.21.0
trio-websocket==0.9.2
typed-ast==1.4.3
types-python-dateutil==2.8.19
types-requests==2.28.8
types-setuptools==64.0.1
types-toml==0.10.8
types-urllib3==1.26.22
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.11
values==2020.12.3
vectorbt==0.23.3
virtualenv==20.16.3
virtualenv-clone==0.5.7
wcwidth==0.2.5
webdriver-manager==3.7.0
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==2.2.2
widgetsnbextension==3.6.0
wrapt==1.14.1
wsproto==1.1.0
yarl==1.7.2
yfinance==0.1.74
zipp==3.8.1
zope.interface==5.4.0
